Added date picker helpers
